pyproject.toml
[tool.poetry]
name = "fohow"
version = "0.1.0"
description = ""
authors = ["It Soda <nikshchegilskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djoser = "^2.2.2"
pillow = "9.2.0"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.2.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
mysqlclient = "^2.2.0"
black = "23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
django-redis = "^5.4.0"
coverage = "^7.3.2"
yookassa = "2.4.0"
pytelegrambotapi = "4.13.0"
psycopg2-binary = "2.9.3"
celery = "5.3.4"
redis = "4.6.0"
drf-extensions = "0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
